feat: switch some things from size_t to INT64 to support george's use of -1 as a logic flag for a number of data structure ops

+++ Alchemy/DirectXUtil/AGScreen.cpp
@@ -72,7 +72,7 @@ void AGScreen::DestroyArea (AGArea *pArea)
 
 	RECT rcRect = pArea->GetRect();
 
-	size_t iIndex;
+	INT64 iIndex;
 	if (m_Areas.Find(pArea, &iIndex))
 		{
 		m_Areas.Delete(iIndex);
